refactor(catalogs): extract Tower's DuckDB usage into a tower-duckdb crate and wire the sandbox in
The sandbox primitives from the previous commit had no caller: the gates,
the session hardening, and the row cap all lived in an inline `#[allow(dead_code)]`
module in catalogs.rs, proven only by tests. This moves Tower's DuckDB usage
into a dedicated `tower-duckdb` crate and makes `tower catalogs query` its first
real consumer, so the hardening runs in production rather than sitting on a shelf.

The crate owns the whole DuckDB surface: opening an in-memory Session, running
trusted setup, locking the session down for untrusted SQL (Session::harden), and
executing a query into JSON rows with an optional row cap. The static text gates
(reject write/DDL, reject multi-statement) live in its `guard` module. The value
conversion, the query execution, and the full adversarial + integration test
suite move with it, so the security invariants are tested next to the code they
protect. tower-cmd no longer depends on duckdb directly; it goes through the crate.

`catalogs query` now splits by mode. Read mode is the default and the sandboxed
path: the SQL is gated before it reaches DuckDB (a smuggled second statement
would otherwise run as a side effect of prepare, and a write gets a clear message
instead of a raw engine error), the session is hardened after attach, and rows
are capped with a truncation notice. Write mode stays the trusted power-user
opt-in and runs as before. Once this lands, #328 will layer the MCP
`tower_catalogs_query` tool on top of the same crate.
