+++ Cargo.toml
@@ -68,6 +68,7 @@ toml = "0.8"
 toml_edit = "0.22"
 tower-api = { path = "crates/tower-api" }
 tower-cmd = { path = "crates/tower-cmd" }
+tower-duckdb = { path = "crates/tower-duckdb" }
 tower-package = { path = "crates/tower-package" }
 tower-runtime = { path = "crates/tower-runtime" }
 tower-telemetry = { path = "crates/tower-telemetry" }
