+++ crates/tower-cmd/Cargo.toml
@@ -11,7 +11,6 @@ clap = { workspace = true }
 cli-table = { workspace = true }
 colored = { workspace = true }
 config = { workspace = true }
-duckdb = { workspace = true }
 crypto = { workspace = true }
 ctrlc = { workspace = true }
 futures-util = { workspace = true }
@@ -30,6 +29,7 @@ spinners = { workspace = true }
 tokio = { workspace = true }
 tokio-util = { workspace = true }
 tower-api = { workspace = true }
+tower-duckdb = { workspace = true }
 tower-package = { workspace = true }
 tower-runtime = { workspace = true }
 tower-telemetry = { workspace = true }
@@ -49,4 +49,3 @@ uuid = { version = "1.0", features = ["v4"] }
 futures = { workspace = true }
 cucumber = { version = "0.21", features = ["macros"] }
 tokio-test = "0.4"
-testcontainers = { version = "0.24", features = ["blocking"] }
