+++ crates/tower-duckdb/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "tower-duckdb"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+chrono = { workspace = true }
+duckdb = { workspace = true }
+serde_json = { workspace = true }
+tower-telemetry = { workspace = true }
+
+[dev-dependencies]
+testcontainers = { version = "0.24", features = ["blocking"] }
